Day 6 task

+++ ecommerceProj/src/main/java/com/ecommerce/repository/UserRepository.java
@@ -1,11 +1,10 @@
 package com.ecommerce.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import com.ecommerce.entity.User;
-
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByEmail(String email);
 }
